requirements.txt
Flask>=3.1,<4
openai>=1.68,<2
python-dotenv>=1.0,<2
openai>=1.68,<2
python-dotenv>=1,<2